+++ sdk/cosmos/azure_data_cosmos/src/models/mod.rs
@@ -12,12 +12,14 @@ use azure_core::{
 };
 use serde::{de::DeserializeOwned, Deserialize, Deserializer, Serialize};
 
+mod connection_string;
 mod container_properties;
 mod indexing_policy;
 mod partition_key_definition;
 mod patch_operations;
 mod throughput_properties;
 
+pub use connection_string::*;
 pub use container_properties::*;
 pub use indexing_policy::*;
 pub use partition_key_definition::*;